Strip Github tar file prefixes in a stream
Previously, the tar file prefixes would be stripped in one go, which
requires filling up an entire buffer, regardless of whether it is
in-memory or file backed. The implementation has now been changed to
use a pipe, so tar file prefixes are stripped lazily when the `.Read`
method is called on a `tarPrefixStripper`.